+++ Inc/audio/firstOrderIirFilter.h
@@ -4,23 +4,23 @@
 
 typedef struct 
 {
-    int16_t oldVal;
-    int16_t oldXVal;
-    int16_t alpha;
+    float oldVal;
+    float oldXVal;
+    float alpha;
 } FirstOrderIirType;
 
 
 typedef struct 
 {
-    int16_t oldVal;
-    int16_t oldXVal;
-    int16_t alphaRising;
-    int16_t alphaFalling;
+    float oldVal;
+    float oldXVal;
+    float alphaRising;
+    float alphaFalling;
 } FirstOrderIirDualCoeffLPType;
 
-int16_t firstOrderIirLowpassProcessSample(int16_t sampleIn,FirstOrderIirType*data);
-int16_t firstOrderIirHighpassProcessSample(int16_t sampleIn,FirstOrderIirType*data);
-int16_t firstOrderIirDualCoeffLPProcessSample(int16_t sampleIn,FirstOrderIirDualCoeffLPType*data);
+float firstOrderIirLowpassProcessSample(float sampleIn,FirstOrderIirType*data);
+float firstOrderIirHighpassProcessSample(float sampleIn,FirstOrderIirType*data);
+float firstOrderIirDualCoeffLPProcessSample(float sampleIn,FirstOrderIirDualCoeffLPType*data);
 
 void firstOrderIirReset(FirstOrderIirType*data);
 void firstOrderIirDualCoeffLPReset(FirstOrderIirDualCoeffLPType*data);
